First pass at refactoring for plugins
